Correct floating point error with RxSignalUnpack
Edit output of RxSignalUnpack to account for Floating-Point Precision

+++ DbcParserLib/Packer.cs
@@ -105,7 +105,7 @@ namespace DbcParserLib
             }
 
             // Apply scaling
-            return ((double)iVal * signal.Factor + signal.Offset);
+            return ((double)(iVal * (decimal)signal.Factor + (decimal)signal.Offset));
         }
 
         /// <summary>
